+++ Sources/XcodeGenCore/Glob.swift
@@ -101,14 +101,12 @@ public class Glob: Collection {
             Glob.patternsCache[pattern] = patterns
         }
         
-        paths = patterns.parallelMap { pattern -> [String] in
-            var gt = glob_t()
-            defer { globfree(&gt) }
-            if executeGlob(pattern: pattern, gt: &gt) {
-                 return populateFiles(gt: gt, includeFiles: includeFiles)
-            }
-            return []
-        }.flatMap { $0 }
+        #if os(macOS)
+        paths = patterns.parallelMap { paths(usingPattern: $0, includeFiles: includeFiles) }.flatMap { $0 }
+        #else
+        // Parallel invocations of Glob on Linux seems to be causing unexpected crashes
+        paths = patterns.map { paths(usingPattern: $0, includeFiles: includeFiles) }.flatMap { $0 }
+        #endif
 
         paths = Array(Set(paths)).sorted { lhs, rhs in
             lhs.compare(rhs) != ComparisonResult.orderedDescending
@@ -225,6 +223,15 @@ public class Glob: Collection {
         isDirectoryCache.removeAll()
     }
 
+    private func paths(usingPattern pattern: String, includeFiles: Bool) -> [String] {
+        var gt = glob_t()
+        defer { globfree(&gt) }
+        if executeGlob(pattern: pattern, gt: &gt) {
+             return populateFiles(gt: gt, includeFiles: includeFiles)
+        }
+        return []
+    }
+    
     private func populateFiles(gt: glob_t, includeFiles: Bool) -> [String] {
         var paths = [String]()
         let includeDirectories = behavior.includesDirectoriesInResults